// app/types/magang.ts
export type Magang = {
  id: string;
  name: string;
  activity_name: string;
  total: number;
  start_duration: string;
  end_duration: string;
  start_registration: string;
  end_registration: string;
  credits_count: number;
  activity_type: string;
  location: string;
  location_kotakab_code: string;
  mitra_id: string;
  certified: boolean;
  logo: string;
  mitra_name: string;
  published_time: string;
};